+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- HTTP logging support for OAuth 2.0 Password Grant authentication, by @HardNorth
+### Fixed
+- Explicitly disable proxy for OAuth authentication when `rp.oauth.use.proxy=false` to avoid issues when proxy is set through system properties, by @HardNorth
 
 ## [5.4.6]
 ### Added